let forward circuits resolve from pending, default channel-failed peer message, u32 fee base, reserve unset htlc origin and define payment retries (NL-200 / W0-C)
(cherry picked from commit f42ac67c162a9cc0213f9dc593d78772bcda79bb)

+++ src/NLightning.Domain/Client/Responses/ChannelInfoClientResponse.cs
@@ -72,7 +72,7 @@ public sealed class ChannelInfoClientResponse
     /// <summary>
     /// Our forwarding <c>fee_base_msat</c> on this channel (what a route hint through us must use).
     /// </summary>
-    public ulong FeeBaseMsat { get; init; }
+    public uint FeeBaseMsat { get; init; }
 
     /// <summary>
     /// Our forwarding <c>fee_proportional_millionths</c> on this channel.
